+++ example/lib/main.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'dart:async';
 
-import 'package:flutter/services.dart';
 import 'package:card_scanner/card_scanner.dart';
 
 void main() {
@@ -16,27 +15,11 @@ class MyApp extends StatefulWidget {
 class _MyAppState extends State<MyApp> {
   CardDetails _cardDetails;
 
-  @override
-  void initState() {
-    super.initState();
-  }
-
-  // Platform messages are asynchronous, so we initialize in an async method.
   Future<void> scanCard() async {
-    CardDetails cardDetails;
-    try {
-      cardDetails =
-          await CardScanner.scanCard(scanOptions: CardScanOptions(scanCardHolderName: true, scanCardIssuer: true));
-    } on PlatformException catch (e) {
-      print('Failed to get platform version : $e');
-      return;
-    }
+    var cardDetails =
+        await CardScanner.scanCard(scanOptions: CardScanOptions(scanCardHolderName: true, scanCardIssuer: true));
 
-    // If the widget was removed from the tree while the asynchronous platform
-    // message was in flight, we want to discard the reply rather than calling
-    // setState to update our non-existent appearance.
     if (!mounted) return;
-
     setState(() {
       _cardDetails = cardDetails;
     });
@@ -47,7 +30,7 @@ class _MyAppState extends State<MyApp> {
     return MaterialApp(
       home: Scaffold(
         appBar: AppBar(
-          title: const Text('Plugin example app'),
+          title: const Text('card_scanner app'),
         ),
         body: Center(
           child: Column(