+++ test/card_scanner_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:card_scanner/card_scanner.dart';
 
 void main() {
   const MethodChannel channel = MethodChannel('nateshmbhat/card_scanner');