+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 0.0.1
+## 0.1.0
++ Added IOS support 🥳
++ Still Better accuracy and speed of scanning ✅
++ refactored example app 📖
 
+
+### 0.0.4
+* Much Better accuracy in card holder name and expiry date scanning ✅
++ Improved speed of scanning 🏎
+
+
+## 0.0.1
 * First Release